refactor: 💡 components / containers / pages 디렉토리 구분
