+++ mytour-front-app/src/Routes.jsx
@@ -7,7 +7,6 @@ import TourRouteIndex from './components/tourRoute/TourRouteIndex'
 import Signup from './Pages/Auth/Signup';
 //import EmailSignup from './components/auth/EmailSignup'
 import EmailSignupContainer from "./containers/EmailSignupContainer"
-import EmailSignup from './containers/auth/EmailSignup'
 
 import Review from './Pages/Review/Review'
